requirements.txt
python-dotenv==1.0.0
requests==2.31.0
pymupdf==1.26.0
nltk==3.9.1
google-cloud-storage==3.1.0
google-cloud-bigquery==3.33.0
google-generativeai==0.8.5
PyPDF2==3.0.1 
streamlit==1.37.0
httpx==0.27.0
pandas==2.2.2
matplotlib==3.9.0
plotly==5.21.0 